requirements.txt
# Core dependencies
python-dotenv
scipy
pydantic

# Web framework
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy[asyncio]
asyncpg
psycopg2-binary

# LangChain and LangGraph dependencies
langchain
langchain_google_genai
langchain_openai
langchain_groq
langchain-community
langchain-core
langchain-text-splitters
langgraph

# Google AI/Cloud dependencies
google-genai
google-generativeai

# Vector stores
langchain-chroma
chromadb

# Document processing
pypdf

# TTS - Edge TTS (primary) and gTTS (fallback)
edge-tts
gtts

# STT - Whisper (primary) and SpeechRecognition (fallback)
openai-whisper
SpeechRecognition

# Audio processing
soundfile
pydub
numpy

# PDF generation
reportlab

# Visualization (optional)
graphviz
ipython
